+++ src/cleanlab_tlm/utils/vpc/chat_completions.py
@@ -48,6 +48,7 @@ class TLMChatCompletion(BaseTLM):
         *,
         options: Optional[TLMOptions] = None,
         timeout: Optional[float] = None,
+        request_headers: Optional[dict[str, str]] = None,
     ):
         """
         lazydocs: ignore
@@ -62,6 +63,7 @@ class TLMChatCompletion(BaseTLM):
             verbose=False,
             allow_custom_model=True,
         )
+        self._request_headers = request_headers or {}
 
     def score(
         self,
@@ -93,6 +95,7 @@ class TLMChatCompletion(BaseTLM):
                 **openai_kwargs,
             },
             timeout=self._timeout,
+            headers=self._request_headers,
         )
 
         res_json = res.json()